Run optimization on PNG and JPG files.
The aim is to reduce the loading time of the website. All available
files were optimized, not just the ones that are visible on the site
directly, so the space saving listed further down is just FYI.

To find and optimize all JPG files, the following command was run:

    find . -name "*.jpg" -type f -exec jpegtran -copy none -optimize -outfile {} {} \;

The total space of JPG files was reduced from 3388k to 2956k.

To find and optimize all PNG files, the follwing command was run:

    find . -name "*.png" -exec optipng -o5 {}  \;

The total space for PNG files was reduced from 32344k to 30828k, smaller
files (icons) usually having larger percentage of reduction, while
the large photos stored in PNG seen smaller reduction, though the latter
are not used on the site directly, so they don't matter much.